chore(navigation): add remaining modals
