+++ app/src/main/java/com/getcode/navigation/screens/Screens.kt
@@ -27,52 +27,3 @@ sealed interface NamedScreen {
     val hasName: Boolean
         @Composable get() = name.isNotEmpty()
 }
-
-sealed interface ModalContent {
-    @Composable
-    fun ModalContainer(
-        displayLogo: Boolean = false,
-        backButton: (Screen?) -> Boolean = { false },
-        closeButton: (Screen?) -> Boolean = { false },
-        onLogoClicked: () -> Unit = { },
-        screenContent: @Composable () -> Unit
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .fillMaxHeight(sheetHeight)
-        ) {
-            val navigator = LocalCodeNavigator.current
-
-            val name = (navigator.lastItem as? NamedScreen)?.name
-            val sheetName = remember(navigator) { name }
-
-            SheetTitle(
-                modifier = Modifier.padding(horizontal = 20.dp),
-                title = sheetName.takeIf { !displayLogo },
-                displayLogo = displayLogo,
-                onLogoClicked = onLogoClicked,
-                // hide while transitioning to/from other destinations
-                backButton = backButton(navigator.lastItem),
-                closeButton = closeButton(navigator.lastItem),
-                onBackIconClicked = { navigator.pop() },
-                onCloseIconClicked = { navigator.hide() }
-            )
-            Box(modifier = Modifier.windowInsetsPadding(WindowInsets.navigationBars)) {
-                screenContent()
-            }
-        }
-    }
-}
-sealed interface ModalRoot: ModalContent
-
-data object MainRoot : Screen {
-    private fun readResolve(): Any = this
-
-    @Composable
-    override fun Content() {
-        // TODO: potentially add a loading state here
-        //  so app doesn't appear stuck in a dead state
-        //  while we wait for auth check to complete
-    }
-}
